separate drawCanvasPolygons
